requirements.txt
semantic-kernel
python-a2a
fastapi
uvicorn
a2a-sdk
python-multipart
python-dotenv